Update README.md
The Project PBL (Problem-Based Learning) on Stock Market Analysis is centered on equipping participants with the skills necessary to assess the viability of potential investments in the stock market. It involves a comprehensive exploration of company fundamentals, enabling individuals to conduct in-depth analyses of financial statements and economic indicators. Through this rigorous process, participants gain the ability to make well-informed investment decisions.